updated ArticleCard
