+++ packages/react-app/src/components/LatestArticles.jsx
@@ -5,7 +5,7 @@ import "slick-carousel/slick/slick.css";
 import favImage from "../assets/favourite.png";
 import nextImage from "../assets/next.png";
 import prevImage from "../assets/prev.png";
-import { ArticleCard } from "../components";
+import { LatestArticleCard } from "../components";
 
 class LatestArticles extends Component {
   constructor(props) {
@@ -87,7 +87,7 @@ class LatestArticles extends Component {
         <div className="relative roadmap_container">
           <Slider ref={c => (this.slider = c)} {...settings}>
             {this.props.articles.map((article_id, index) => (
-              <ArticleCard key={index} id={article_id}></ArticleCard>
+              <LatestArticleCard key={index} id={article_id}></LatestArticleCard>
             ))}
           </Slider>
           <div className="md:hidden absolute top-1/2 left-4">
