+++ packages/react-app/src/components/UserArticles.jsx
@@ -1,7 +1,7 @@
 import { useEffect, useState } from "react";
 import { useNavigate } from "react-router-dom";
 import illustrationImage from "../assets/illustration.png";
-import { ArticleCard } from "../components";
+import { LatestArticleCard } from "../components";
 
 const UserArticles = ({ address }) => {
   const [articles, setArticles] = useState([]);
@@ -35,7 +35,7 @@ const UserArticles = ({ address }) => {
       ) : (
         <div className="grid grid-cols-1 md:grid-cols-2 lg:grid-cols-3 xl:grid-cols-4 gap-4">
           {articles.map((item, index) => {
-            return <ArticleCard article={item}></ArticleCard>;
+            return <LatestArticleCard article={item}></LatestArticleCard>;
           })}
         </div>
       )}
