+++ packages/react-app/src/views/AdvancedSearch.jsx
@@ -2,7 +2,7 @@ import axios from "axios";
 import React, { useEffect, useState } from "react";
 import clear from "../assets/clear.svg";
 import search from "../assets/search.svg";
-import { AuthorCard, Footer, SubmissionCard } from "../components";
+import { AuthorCard, Footer, ArticleCard } from "../components";
 import { strcmp } from "../utils/utils";
 
 const server = "https://tdao-api.herokuapp.com";
@@ -100,7 +100,7 @@ const AdvancedSearch = () => {
   };
 
   return (
-    <div className="flex flex-col" style={{ backgroundImage: "linear-gradient(#fff, #EEEE" }}>
+    <div style={{ backgroundImage: "linear-gradient(#fff, #EEEE" }}>
       <div className="relative" style={{ backgroundColor: "#e2e2e2" }}>
         <div className="lg:mx-auto lg:max-w-2xl overflow-hidden relative text-left space-y-8 py-16">
           <div className="mx-4 flex flex-col items-center justify-center space-y-8">
@@ -109,15 +109,16 @@ const AdvancedSearch = () => {
                 type="text"
                 className="w-full px-4 text-black text-lg focus:outline-none"
                 value={value}
+                placeholder="Input Keyword"
                 onChange={e => setValue(e.target.value)}
                 onKeyDown={handleKeyDown}
               ></input>
               <div
-                className="w-40 bg-primary rounded-full py-2 text-sm flex flex-row items-center justify-center cursor-pointer"
+                className="w-12 lg:w-40 bg-primary rounded-full py-2 text-sm flex flex-row items-center justify-center cursor-pointer space-x-2"
                 onClick={handleSearch}
               >
                 <img className="" src={search} width={24} height={24}></img>
-                <div>Search</div>
+                <div className="hidden lg:inline">Search</div>
               </div>
             </div>
             <div className="w-full flex flex-col md:flex-row items-start md:items-end space-x-0 md:space-x-8 space-y-4 md:space-y-0">
@@ -126,7 +127,7 @@ const AdvancedSearch = () => {
                 <select
                   id="select-blockchain"
                   name="select-blockchain"
-                  className="mt-1 block bg-transparent w-full pl-3 pr-10 py-2 text-lg rounded-xl border border-black"
+                  className="mt-1 block bg-white w-full pl-3 pr-10 py-2 text-lg rounded-lg"
                   value={sortField}
                   onChange={handleSortFieldChange}
                 >
@@ -157,7 +158,7 @@ const AdvancedSearch = () => {
                 <select
                   id="select-blockchain"
                   name="select-blockchain"
-                  className="mt-1 block bg-transparent w-full pl-3 pr-10 py-2 text-lg rounded-xl border border-black"
+                  className="mt-1 block bg-white w-full pl-3 pr-10 py-2 text-lg rounded-lg"
                   value={category}
                   onChange={handleCategoryChange}
                 >
@@ -170,7 +171,7 @@ const AdvancedSearch = () => {
                 <select
                   id="select-blockchain"
                   name="select-blockchain"
-                  className="mt-1 block bg-transparent w-full pl-3 pr-10 py-2 text-lg rounded-xl border border-black"
+                  className="mt-1 block bg-white w-full pl-3 pr-10 py-2 text-lg rounded-lg"
                 >
                   <option>English</option>
                 </select>
@@ -201,7 +202,7 @@ const AdvancedSearch = () => {
           </div>
         </div>
       </div>
-      <div className="px-4 sm:px-8 md:px-10 xl:px-20 overflow-hidden">
+      <div className="mx-auto pt-4 max-w-xl md:max-w-4xl xl:max-w-7xl">
         <div className="text-sm pt-8 text-left">
           {category === "author" ? (
             <div>{searchResult.length} similar Authors found</div>
@@ -209,14 +210,12 @@ const AdvancedSearch = () => {
             <div>{searchResult.length} similar Articles found</div>
           )}
         </div>
-        <div className="py-4 grid grid-cols-1 md:grid-cols-2 lg:grid-cols-3 xl:grid-cols-4 gap-4">
+        <div className="py-4 grid grid-cols-1 md:grid-cols-2 lg:grid-cols-3 xl:grid-cols-4 gap-8">
           {category === "author"
             ? searchResult.map(item => <AuthorCard key={Math.random()} author={item}></AuthorCard>)
-            : searchResult.map(item => <SubmissionCard key={Math.random()} article={item}></SubmissionCard>)}
+            : searchResult.map(item => <ArticleCard key={Math.random()} article={item}></ArticleCard>)}
         </div>
-      </div>
-      <div className="px-4 sm:px-8 md:px-10 xl:px-20">
-        <Footer></Footer>
+        <Footer />
       </div>
     </div>
   );
